[feat]: update v1.5.7
* [build]: updated dependencies

* [fix]: improved code around the whole project, fixed #106 and #107

* [feat]: session cache, closes #103

* [feat]: update zh_rCN translation

---------

+++ app/src/main/java/com/f0x1d/logfox/repository/logging/RecordingsRepository.kt
@@ -8,21 +8,24 @@ import com.f0x1d.logfox.extensions.context.toast
 import com.f0x1d.logfox.extensions.notifications.cancelRecordingNotification
 import com.f0x1d.logfox.extensions.notifications.sendRecordingNotification
 import com.f0x1d.logfox.extensions.notifications.sendRecordingPausedNotification
-import com.f0x1d.logfox.extensions.onAppScope
-import com.f0x1d.logfox.extensions.runOnAppScope
 import com.f0x1d.logfox.model.LogLine
 import com.f0x1d.logfox.repository.logging.base.LoggingHelperItemsRepository
 import com.f0x1d.logfox.repository.logging.readers.recordings.RecordingWithFiltersReader
+import com.f0x1d.logfox.repository.logging.readers.recordings.RewritingRecordingReader
 import com.f0x1d.logfox.utils.DateTimeFormatter
 import com.f0x1d.logfox.utils.preferences.AppPreferences
 import com.f0x1d.logfox.utils.terminal.base.Terminal
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.job
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.io.File
 import java.io.FileOutputStream
@@ -35,6 +38,7 @@ class RecordingsRepository @Inject constructor(
     private val database: AppDatabase,
     private val dateTimeFormatter: DateTimeFormatter,
     private val recordingReader: RecordingWithFiltersReader,
+    private val cacheRecordingReader: RewritingRecordingReader,
     private val appPreferences: AppPreferences,
     private val terminals: Array<Terminal>
 ): LoggingHelperItemsRepository<LogRecording>() {
@@ -42,44 +46,87 @@ class RecordingsRepository @Inject constructor(
     val recordingStateFlow = MutableStateFlow(RecordingState.IDLE)
 
     override val readers = listOf(
-        recordingReader
+        recordingReader,
+        cacheRecordingReader
     )
 
-    private val recordingDir = File("${context.filesDir.absolutePath}/recordings").apply {
+    private val recordingsDir = File("${context.filesDir.absolutePath}/recordings").apply {
+        if (!exists()) mkdirs()
+    }
+    private val cacheRecordingsDir = File("${context.filesDir.absolutePath}/recordings/cache").apply {
         if (!exists()) mkdirs()
     }
 
-    private var filtersJob: Job? = null
+    private var useLogsCache = appPreferences.useSessionCache
+    private var savingCacheRecording = appPreferences.saveSessionCacheToRecordings
 
     override suspend fun setup() {
-        filtersJob = onAppScope {
-            database.userFilterDao().getAllAsFlow()
-                .distinctUntilChanged()
-                .flowOn(Dispatchers.IO)
-                .collect {
-                    recordingReader.updateFilters(it)
-                }
-        }
+        super.setup()
 
-        File(recordingDir, "all.log").delete()
+        database.userFilterDao().getAllAsFlow()
+            .flowOn(Dispatchers.IO)
+            .onEach {
+                recordingReader.updateFilters(it)
+                cacheRecordingReader.updateFilters(it)
+            }
+            .launchIn(repositoryScope)
+
+        // Deprecated
+        File(recordingsDir, "all.log").delete()
+
+        // Updated only in setup to avoid saving deleted file and etc.
+        useLogsCache = appPreferences.useSessionCache
+        savingCacheRecording = appPreferences.saveSessionCacheToRecordings
+
+        if (useLogsCache) {
+            cacheRecordingReader.record(
+                File(
+                    cacheRecordingsDir,
+                    "${dateTimeFormatter.formatForExport(System.currentTimeMillis())}.log"
+                )
+            )
+
+            repositoryScope.launch {
+                while (isActive) {
+                    delay(1000)
+                    cacheRecordingReader.dumpLines()
+                }
+            }
+
+            if (savingCacheRecording) database.logRecordingDao().insert(
+                LogRecording(
+                    title = "${context.getString(R.string.session_cache)} ${database.logRecordingDao().count(cached = true) + 1}",
+                    dateAndTime = cacheRecordingReader.recordingTime,
+                    file = cacheRecordingReader.recordingFile ?: return,
+                    isCacheRecording = true
+                )
+            )
+        }
     }
 
     override suspend fun stop() {
-        if (recordingStateFlow.value != RecordingState.IDLE) {
-            recordingReader.updateRecording(false)
-            recordingReader.deleteFile()
+        if (useLogsCache) {
+            cacheRecordingReader.stopRecording()
+            if (!savingCacheRecording)
+                cacheRecordingReader.recordingFile?.delete()
         }
 
-        recordingStateFlow.update { RecordingState.IDLE }
-        recordingReader.clearLines()
+        when (recordingStateFlow.value) {
+            RecordingState.RECORDING,
+            RecordingState.PAUSED -> end(recordingSaved = {}).join()
 
-        filtersJob?.cancel()
+            RecordingState.SAVING -> repositoryScope.coroutineContext.job.join()
+
+            else -> {}
+        }
+
+        super.stop()
     }
 
-    fun saveAll(recordingSaved: (LogRecording) -> Unit = {}) = runOnAppScope {
+    fun saveAll(recordingSaved: (LogRecording) -> Unit = {}) = runOnRepoScope {
         val recordingTime = System.currentTimeMillis()
         val recordingFile = File(
-            recordingDir,
+            recordingsDir,
             "${dateTimeFormatter.formatForExport(recordingTime)}.log"
         )
 
@@ -102,9 +149,9 @@ class RecordingsRepository @Inject constructor(
         }
 
         val logRecording = LogRecording(
-            "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}",
-            recordingTime,
-            recordingFile.absolutePath
+            title = "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}",
+            dateAndTime = recordingTime,
+            file = recordingFile
         ).let {
             it.copy(id = database.logRecordingDao().insert(it))
         }
@@ -114,11 +161,11 @@ class RecordingsRepository @Inject constructor(
         }
     }
 
-    fun createRecordingFrom(lines: List<LogLine>) = runOnAppScope {
+    fun createRecordingFrom(lines: List<LogLine>) = runOnRepoScope {
         val recordingTime = System.currentTimeMillis()
 
         val recordingFile = File(
-            recordingDir,
+            recordingsDir,
             "${dateTimeFormatter.formatForExport(recordingTime)}.log"
         )
 
@@ -128,23 +175,21 @@ class RecordingsRepository @Inject constructor(
             }
         )
 
-        val title = "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}"
-
         database.logRecordingDao().insert(
             LogRecording(
-                title,
-                recordingTime,
-                recordingFile.absolutePath
+                title = "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}",
+                dateAndTime = recordingTime,
+                file = recordingFile
             )
         )
     }
 
-    fun record() = runOnAppScope {
+    fun record() = runOnRepoScope {
         recordingStateFlow.update { RecordingState.RECORDING }
 
         recordingReader.record(
             File(
-                recordingDir,
+                recordingsDir,
                 "${dateTimeFormatter.formatForExport(System.currentTimeMillis())}.log"
             )
         )
@@ -152,31 +197,27 @@ class RecordingsRepository @Inject constructor(
         context.sendRecordingNotification()
     }
 
-    fun pause() = runOnAppScope {
+    fun pause() = runOnRepoScope {
         recordingStateFlow.update { RecordingState.PAUSED }
         recordingReader.updateRecording(false)
         context.sendRecordingPausedNotification()
     }
 
-    fun resume() = runOnAppScope {
+    fun resume() = runOnRepoScope {
         recordingStateFlow.update { RecordingState.RECORDING }
         recordingReader.updateRecording(true)
         context.sendRecordingNotification()
     }
 
-    fun end(recordingSaved: (LogRecording) -> Unit = {}) = runOnAppScope {
+    fun end(recordingSaved: (LogRecording) -> Unit = {}) = onRepoScope {
         recordingStateFlow.update { RecordingState.SAVING }
-        recordingReader.updateRecording(false)
+        recordingReader.stopRecording()
         context.cancelRecordingNotification()
 
-        recordingReader.dumpLines()
-
-        val title = "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}"
-
         val logRecording = LogRecording(
-            title,
-            recordingReader.recordingTime,
-            recordingReader.recordingFile?.absolutePath ?: return@runOnAppScope
+            title = "${context.getString(R.string.record_file)} ${database.logRecordingDao().count() + 1}",
+            dateAndTime = recordingReader.recordingTime,
+            file = recordingReader.recordingFile ?: return@onRepoScope
         ).let {
             it.copy(id = database.logRecordingDao().insert(it))
         }
@@ -190,10 +231,25 @@ class RecordingsRepository @Inject constructor(
 
     fun updateTitle(logRecording: LogRecording, newTitle: String) = update(logRecording.copy(title = newTitle))
 
+    fun clearCached() = runOnRepoScope {
+        database.logRecordingDao().getAll(cached = true).forEach {
+            if (it.file != cacheRecordingReader.recordingFile)
+                it.deleteFile()
+            else
+                // To delete it later
+                savingCacheRecording = false
+        }
+        database.logRecordingDao().deleteAll(deleteCacheRecordings = true)
+    }
+
     override suspend fun updateInternal(item: LogRecording) = database.logRecordingDao().update(item)
 
     override suspend fun deleteInternal(item: LogRecording) {
-        item.deleteFile()
+        if (item.file != cacheRecordingReader.recordingFile)
+            item.deleteFile()
+        else
+            // To delete it later
+            savingCacheRecording = false
         database.logRecordingDao().delete(item)
     }
